# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ResponseError:
    properties:
      errors:
        example: failed get data
        type: string
      message:
        example: Failed get data
        type: string
      status:
        example: Error
        type: string
    type: object
  dto.ResponseSuccess:
    properties:
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
    type: object
...
  dto.UpdatePasswordRequest:
    properties:
      new_password:
        example: example321
        minLength: 8
        type: string
      old_password:
        example: example123
        minLength: 8
        type: string
    required:
      - new_password
      - old_password
    type: object
...
  dto.User:
    properties:
      address:
        example: Jakarta
        type: string
      created_at:
        example: "2025-01-01T09:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      fullname:
        example: John Doe
        type: string
      id:
        example: 1
        type: integer
      last_login:
        example: "2025-01-01T10:00:00Z"
        type: string
      phone:
        example: "081234567890"
        type: string
      photo:
        example: avatar.png
        type: string
      role:
        example: user
        type: string
    type: object
  dto.UserProfileResponse:
    properties:
      data:
        $ref: "#/definitions/dto.User"
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
    type: object
host: 192.168.50.221:8080
info:
  contact: {}
  description: Solid Coffee Backend RESTful API
  title: Solid Coffee Backend
  version: "1.0"
paths:
...
  /admin/user/:
    get:
      description: Get authenticated user's profile information
      parameters:
        - description: Page number
          in: query
          name: page
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.UserProfileResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Get all user profile
      tags:
        - Admin User Management
    post:
      consumes:
        - multipart/form-data
      description: Insert authenticated user's profile information
      parameters:
        - description: Profile photo
          in: formData
          name: photo
          required: true
          type: file
        - description: Full name (min 3 chars)
          in: formData
          name: fullname
          required: true
          type: string
        - description: Email
          in: formData
          name: email
          required: true
          type: string
        - description: Password (min 8 chars)
          in: formData
          name: password
          required: true
          type: string
        - description: Phone number (min 3 chars)
          in: formData
          name: phone
          required: true
          type: string
        - description: Address (min 3 chars)
          in: formData
          name: address
          required: true
          type: string
        - description: role (user or admin)
          in: formData
          name: role
          required: true
          type: string
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Insert user profile
      tags:
        - Admin User Management
  /admin/user/{id}:
    delete:
      consumes:
        - multipart/form-data
      description: Delete authenticated user's profile information
      parameters:
        - description: user id
          in: path
          name: id
          type: integer
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Delete user profile
      tags:
        - Admin User Management
    patch:
      consumes:
        - multipart/form-data
      description: Update authenticated user's profile information
      parameters:
        - description: user id
          in: path
          name: id
          type: integer
        - description: Profile photo
          in: formData
          name: photo
          type: file
        - description: Full name (min 3 chars)
          in: formData
          name: fullname
          type: string
        - description: Phone number (min 3 chars)
          in: formData
          name: phone
          type: string
        - description: Address (min 3 chars)
          in: formData
          name: address
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Update user profile
      tags:
        - Admin User Management
...
  /user/:
    get:
      description: Get authenticated user's profile information
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.UserProfileResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Get user profile
      tags:
        - Users
    patch:
      consumes:
        - multipart/form-data
      description: Update authenticated user's profile information
      parameters:
        - description: Profile photo
          in: formData
          name: photo
          type: file
        - description: Full name (min 3 chars)
          in: formData
          name: fullname
          type: string
        - description: Phone number (min 3 chars)
          in: formData
          name: phone
          type: string
        - description: Address (min 3 chars)
          in: formData
          name: address
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Update user profile
      tags:
        - Users
  /user/password/:
    patch:
      consumes:
        - application/json
      description: Update authenticated user's password
      parameters:
        - description: Edit password data
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.UpdatePasswordRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Change user password
      tags:
        - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"